+++ Memoria/Frame.h
@@ -9,7 +9,7 @@
 // todo: empaquetar esta estructura?
 typedef struct
 {
-    uint32_t Timestamp;
+    uint64_t Timestamp;
     uint16_t Key;
     char Value[]; // tamaño maximo seteado en tiempo de ejecucion!
 } Frame;
@@ -17,7 +17,7 @@ typedef struct
 typedef struct
 {
     char const* TableName;
-    uint32_t Timestamp;
+    uint64_t Timestamp;
     uint16_t Key;
     char const* Value;
 } DirtyFrame;
